# [numpy] FFI for insert \ delete \ matmul etc.