--- a/python/mxnet/ndarray/numpy/random.py
+++ b/python/mxnet/ndarray/numpy/random.py
@@ -180,21 +180,14 @@ def normal(loc=0.0, scale=1.0, size=None, dtype=None, ctx=None, out=None):
     if dtype is None:
         dtype = 'float32'
     if ctx is None:
-        ctx = current_context()
+        ctx = str(current_context())
+    else:
+        ctx = str(ctx)
+    if dtype is not None and not isinstance(dtype, str):
+        dtype = _np.dtype(dtype).name
     if size == ():
         size = None
-    if input_type == (True, True):
-        return _npi.normal(loc, scale, loc=None, scale=None, size=size,
-                           ctx=ctx, dtype=dtype, out=out)
-    elif input_type == (False, True):
-        return _npi.normal(scale, loc=loc, scale=None, size=size,
-                           ctx=ctx, dtype=dtype, out=out)
-    elif input_type == (True, False):
-        return _npi.normal(loc, loc=None, scale=scale, size=size,
-                           ctx=ctx, dtype=dtype, out=out)
-    else:
-        return _npi.normal(loc=loc, scale=scale, size=size,
-                           ctx=ctx, dtype=dtype, out=out)
+    return _api_internal.normal(loc, scale, size, ctx, dtype, out)
 
 
 def lognormal(mean=0.0, sigma=1.0, size=None, dtype=None, ctx=None, out=None):
@@ -802,18 +795,6 @@ def gamma(shape, scale=1.0, size=None, dtype=None, ctx=None, out=None):
     if dtype is not None and not isinstance(dtype, str):
         dtype = _np.dtype(dtype).name
     return _api_internal.gamma(shape, scale, size, ctx, dtype, out)
-    # if input_type == (True, True):
-    #     return _npi.gamma(shape, scale, shape=None, scale=None, size=size,
-    #                       ctx=ctx, dtype=dtype, out=out)
-    # elif input_type == (False, True):
-    #     return _npi.gamma(scale, shape=shape, scale=None, size=size,
-    #                       ctx=ctx, dtype=dtype, out=out)
-    # elif input_type == (True, False):
-    #     return _npi.gamma(shape, shape=None, scale=scale, size=size,
-    #                       ctx=ctx, dtype=dtype, out=out)
-    # else:
-    #     return _npi.gamma(shape=shape, scale=scale, size=size,
-    #                       ctx=ctx, dtype=dtype, out=out)
 
     raise ValueError("Distribution parameters must be either mxnet.numpy.ndarray or numbers")