--- a/src/api/operator/numpy/np_delete_op.cc
+++ b/src/api/operator/numpy/np_delete_op.cc
@@ -21,6 +21,7 @@
  * \file np_delete_op.cc
  * \brief Implementation of the API of functions in src/operator/numpy/np_delete_op.cc
  */
+#include <vector>
 #include "../utils.h"
 #include "../../../operator/numpy/np_delete_op-inl.h"
 
@@ -29,11 +30,10 @@ namespace mxnet {
 MXNET_REGISTER_API("_npi.delete")
 .set_body([](runtime::MXNetArgs args, runtime::MXNetRetValue* ret) {
   using namespace runtime;
-  const nnvm::Op* op = Op::Get("_npi_delete");
+  static const nnvm::Op* op = Op::Get("_npi_delete");
   nnvm::NodeAttrs attrs;
   op::NumpyDeleteParam param;
-  NDArray** inputs = NULL;
-  int num_inputs;
+  int num_inputs = 0;
   param.start = dmlc::nullopt;
   param.step = dmlc::nullopt;
   param.stop = dmlc::nullopt;
@@ -43,7 +43,6 @@ MXNET_REGISTER_API("_npi.delete")
     if (args[1].type_code() == kDLInt ||
         args[1].type_code() == kDLUInt ||
         args[1].type_code() == kDLFloat) {
-      inputs = new NDArray*[1]{args[0].operator mxnet::NDArray*()};
       if (args[1].type_code() == kNull) {
         param.int_ind = dmlc::nullopt;
       } else {
@@ -56,8 +55,6 @@ MXNET_REGISTER_API("_npi.delete")
       }
       num_inputs = 1;
     } else {
-      inputs = new NDArray*[2]{args[0].operator mxnet::NDArray*(),
-                               args[1].operator mxnet::NDArray*()};
       if (args[2].type_code() == kNull) {
         param.axis = dmlc::nullopt;
       } else {
@@ -66,7 +63,6 @@ MXNET_REGISTER_API("_npi.delete")
       num_inputs = 2;
     }
   } else {
-    inputs = new NDArray*[1]{args[0].operator mxnet::NDArray*()};
     num_inputs = 1;
     if (args[1].type_code() == kNull) {
       param.start = dmlc::nullopt;
@@ -89,10 +85,15 @@ MXNET_REGISTER_API("_npi.delete")
       param.axis = args[4].operator int();
     }
   }
+  std::vector<NDArray*> inputs;
+  for (int i = 0; i < num_inputs; ++i) {
+    inputs.push_back(args[i].operator mxnet::NDArray*());
+  }
   attrs.parsed = std::move(param);
   attrs.op = op;
   int num_outputs = 0;
-  auto ndoutputs = Invoke<op::NumpyDeleteParam>(op, &attrs, num_inputs, inputs, &num_outputs, nullptr);
+  auto ndoutputs = Invoke<op::NumpyDeleteParam>
+    (op, &attrs, num_inputs, inputs.data(), &num_outputs, nullptr);
   *ret = ndoutputs[0];
 });