## Description ##
ffi op list:
| op | Old FFI (ctypes) (us) | New FFI (cython) (us) |
| :-: | :-------------------: | :---------------------: |
| append | 31.3 | 7.73 |
| orange | 48.97 | 12.93 |
| blackman | 47.84 | 33.94 |
| concatenate | 34.3 | 8.84 |
| delete | 36.61 | 8.26 |
| eye | 50.8 | 13.13 |
| hamming | 44.78 | 11.95 |
| hanning | 44.69 | 11.82 |
| insert | 31.84 | 8.13 |
| append | 31.3 | 7.73 |
| linspace | 40.8 | 8.68 |
| logspace | 57.78 | 13.34 |
| matmul | 25.11 | 10.61 |
| mean | 33.33 | 21.71 |
| random.gamma | 69.77 | 14.74 |
| random.normal | 67.61 | 15.06 |
| random.uniform | 59.54 | 13.18 |
| append | 31.3 | 7.73 |

## Checklist ##
### Essentials ###
Please feel free to remove inapplicable items for your PR.
- [ ] The PR title starts with [MXNET-$JIRA_ID], where $JIRA_ID refers to the relevant [JIRA issue](https://issues.apache.org/jira/projects/MXNET/issues) created (except PRs with tiny changes)
- [ ] Changes are complete (i.e. I finished coding on this PR)
- [ ] All changes have test coverage:
- Unit tests are added for small changes to verify correctness (e.g. adding a new operator)
- Nightly tests are added for complicated/long-running ones (e.g. changing distributed kvstore)
- Build tests will be added for build configuration changes (e.g. adding a new build option with NCCL)
- [ ] Code is well-documented: 
- For user-facing API changes, API doc string has been updated. 
- For new C++ functions in header files, their functionalities and arguments are documented. 
- For new examples, README.md is added to explain the what the example does, the source of the dataset, expected performance on test set and reference to the original paper if applicable
- Check the API doc at https://mxnet-ci-doc.s3-accelerate.dualstack.amazonaws.com/PR-$PR_ID/$BUILD_ID/index.html
- [ ] To the best of my knowledge, examples are either not affected by this change, or have been fixed to be compatible with this change

### Changes ###
- [ ] Feature1, tests, (and when applicable, API doc)
- [ ] Feature2, tests, (and when applicable, API doc)

## Comments ##
- If this change is a backward incompatible change, why must this change be made.
- Interesting edge cases to note here