@JiangZhaoh 2020-04-08
run ci [centos-cpu]